+++ Jobs/Tank.py
@@ -22,7 +22,7 @@ class Tank(Player):
     def createRecord(
         self,
         target: Player,
-        value: int = 0,
+        value: float = 0,
         effect: list[Effect] | Effect = [],
     ) -> Record:
         return Record(
